Refactor never loop to its own module
